+++ radcad/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.11.0"
+__version__ = "0.11.1"
 
 from radcad.wrappers import Context, Model, Simulation, Experiment
 from radcad.engine import Engine