+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.1] - 2022-09-03
+### Added
+- Add support for parameters dataclass
+- Add support for single value parameters (i.e. not a list)
+
 ## [0.11.0] - 2022-09-03
 ### Changed
 - Adapted core deepcopy processes to be more efficient and avoid unintended mutation of state between policy and state update functions